+++ R/salesforcer.R
@@ -17,5 +17,6 @@ NULL
 if(getRversion() >= "2.15.1")  utils::globalVariables(c(".", 
                                                         "id", 
                                                         "success", 
+                                                        "sObjectType",
                                                         "attributes.type", 
                                                         "sf:type"))